Rename ccache_utils to cache_utils
